refactor(fe/mail): state에서 url로 상태관리 기존코드 걷어내기
