+++ web/pages/index.js
@@ -5,7 +5,6 @@ import MailArea from '../components/MailArea';
 import WriteMailToMe from '../components/WriteMailToMe';
 import WriteMail from '../components/WriteMail';
 import ReadMail from '../components/ReadMail';
-import HeadTitle from '../components/HeadTitle';
 
 const VIEW_COMPONENT = {
   'WRITE-TO-ME': <WriteMailToMe />,
@@ -17,6 +16,7 @@ const IndexPage = () => {
   const router = useRouter();
   const { query } = router;
   let { view } = query;
+
   if (view) {
     view = view.toUpperCase();
   }
@@ -24,7 +24,6 @@ const IndexPage = () => {
 
   return (
     <>
-      <HeadTitle title="다잇누 메일서비스" />
       <Layout>{renderView}</Layout>
     </>
   );
