+++ web/utils/url/change-query.js
@@ -6,6 +6,15 @@ const VIEW_STRING = {
   READ: 'read',
 };
 
+const SORTING_CRITERIA = {
+  datedesc: 'sort=datedesc',
+  dateasc: 'sort=dateasc',
+  subjectdesc: 'sort=subjectdesc',
+  subjectasc: 'sort=subjectasc',
+  fromdesc: 'sort=fromdesc',
+  fromasc: 'sort=fromasc',
+};
+
 const changeView = view => {
   changeUrlWithoutRunning({ view });
 };
@@ -14,13 +23,14 @@ const changeCategory = categoryNo => {
   changeUrlWithoutRunning({ category: categoryNo });
 };
 
-const changeUrlWithoutRunning = ({ category, page, view, mailNo }) => {
+const getQueryByOptions = ({ category, page, view, mailNo, sort }) => {
   const queries = [];
-  if (category && category !== 0) {
+
+  if (category > 0) {
     queries.push(`category=${category}`);
   }
 
-  if (page) {
+  if (page > 0) {
     queries.push(`page=${page}`);
   }
 
@@ -32,6 +42,23 @@ const changeUrlWithoutRunning = ({ category, page, view, mailNo }) => {
     queries.push(`mailNo=${mailNo}`);
   }
 
+  if (SORTING_CRITERIA[sort]) {
+    queries.push(SORTING_CRITERIA[sort]);
+  }
+
+  return queries;
+};
+
+/**
+ * @param {Object} options
+ * @param {Number} options.category
+ * @param {Number} options.page
+ * @param {String} options.view
+ * @param {Number} options.mailNo
+ * @param {String} options.sort
+ */
+const changeUrlWithoutRunning = options => {
+  const queries = getQueryByOptions(options);
   if (queries.length === 0) {
     return Router.push('/', '/', { shallow: true });
   }
@@ -39,7 +66,7 @@ const changeUrlWithoutRunning = ({ category, page, view, mailNo }) => {
   const nextUrl = queries.join('&');
   const href = `/?${nextUrl}`;
   const as = href;
-  Router.push(href, as, { shallow: true });
+  return Router.push(href, as, { shallow: true });
 };
 
-export { changeUrlWithoutRunning, changeView, VIEW_STRING, changeCategory };
+export { changeUrlWithoutRunning, getQueryByOptions, changeView, VIEW_STRING, changeCategory };
